Refactor love.graphics to separate Graphics object
