+++ src/level.lua
@@ -1,18 +1,14 @@
--- This file describes our level
-
-level = {}
-
 require 'src.whale'
 require 'src.constants'
 local audio = require 'src.audio'
-local colors = require 'src.colors'
+Graphics = require 'src.graphics'
+
+level = {}
 
 -- LOAD
 local function createBackgroundQuad()
-  -- Create a quad for the background
-  mapWidth = images.bg_mnt1:getWidth() * 2
-  mountainQuad = love.graphics.newQuad(0, 0, mapWidth, 704, 1547, 704)
-  forestQuad = love.graphics.newQuad(0, 0, mapWidth, 423, 1547, 423)
+  mountainQuad = Graphics:backgroundQuad(704)
+  forestQuad = Graphics:backgroundQuad(423)
   images.bg_mnt1:setWrap("repeat")
   images.bg_mnt2:setWrap("repeat")
   images.bg_forest:setWrap("repeat")
@@ -20,20 +16,8 @@ end
 
 local function loadImages()
   img_fn = {"bg_forest", "bg_mnt1", "bg_mnt2", "fg_snow", "ui_hiscore", "lanes", "logo", "obj_log", "obj_snowman", "obj_stone", "obj_tree", "player_idle", "player_run1", "player_run2", "ui_score", "sky", "vall"}
-  images = {}
-  for _, v in ipairs(img_fn) do
-    images[v] = love.graphics.newImage("assets/"..v..".png")
-  end
-  
+  Graphics:loadImages(img_fn)  
   createBackgroundQuad()
-end
-
-local function initFont()
-  fonts = {
-    score = love.graphics.newFont("assets/font_score.otf", 22),
-    high_score = love.graphics.newFont("assets/font_hiscore.otf", 27),
-    instructions = love.graphics.newFont("assets/font_score.otf", 27),
-  }
 end
 
 local function initLanes()
@@ -63,8 +47,8 @@ local function initAudio()
 end
 
 function level.load()
+  Graphics:loadLevelAssets()
   loadImages()
-  initFont()
   initLanes()
   initAudio()
   level.objects = {}
@@ -183,7 +167,7 @@ local function removeObjects()
   end
 end
 
-function level.update(dt)
+function love.update(dt)
   if not isRunning then
     audioSources.yodel_intro:stop()
     audioSources.yodel_loop:stop()
@@ -237,148 +221,15 @@ function level.update(dt)
   end
 end
 
--- DRAW
-local function drawBackground()
-  -- Draw the beautiful sky
-  love.graphics.draw(images.sky, 0, 0)
-  love.graphics.draw(images.bg_mnt1, mountainQuad, 0, 0, 0, 1, 1, bgOffsets.mnt1, 0)
-  love.graphics.draw(images.bg_mnt2, mountainQuad, 0, 0, 0, 1, 1, bgOffsets.mnt2, 0)
-  love.graphics.draw(images.bg_forest, forestQuad, 0, 191, 0, 1, 1, bgOffsets.forest, 0)
-end
-
-local function drawImage(image, x, y)
-  love.graphics.draw(image, x, y)
-end
-
-local function drawForeground()
-  -- Draw at bottom left
-  drawImage(images.fg_snow, 0, 272)
-end
-
-local function drawLanes()
-  drawImage(images.lanes, 0, 130)
-  drawImage(images.vall, 0, 105)
-  drawImage(images.vall, 0, 345)
-end
-
-local function drawObjects()
-  table.sort(level.objects, function(a, b) return a.lane > b.lane end)
-  playerDrawn = true
-  
-  for _, v in ipairs(level.objects) do
-    if v.lane == 3 then
-      love.graphics.draw(images[v.ID], v.x, v.y, 0, 1, 1, offsets[v.ID][1], offsets[v.ID][2])
-    end
-    if (whale.lane == 3) and isRunning then
-      whale.playAnimation()
-      playerDrawn = false
-    end
-  end
-  for _, v in ipairs(level.objects) do
-    if v.lane == 2 then
-      love.graphics.draw(images[v.ID], v.x, v.y, 0, 1, 1, offsets[v.ID][1], offsets[v.ID][2])
-    end
-    if whale.lane == 2 and isRunning then
-      whale.playAnimation()
-      playerDrawn = false
-    end
-  end
-  for _, v in ipairs(level.objects) do
-    if v.lane == 1 then
-      love.graphics.draw(images[v.ID], v.x, v.y, 0, 1, 1, offsets[v.ID][1], offsets[v.ID][2])
-    end
-    if whale.lane == 1 and isRunning then
-      whale.playAnimation()
-      playerDrawn = false
-    end
-  end
-end
-
-local function setColor(color, alpha)
-  if alpha then
-    local color_table = colors[color]
-    table.insert(color_table, alpha)
-    love.graphics.setColor(color_table)
-  else 
-    love.graphics.setColor(colors[color])
-  end
-end
-
-local function drawGUI(controls)
-  local center = {
-    x = GAME_WIDTH / 2,
-    y = GAME_HEIGHT / 2
-  }
-
-  -- Score and hiscore
-  local limit = 200 -- Wrap the line after this many horizontal pixels
-  drawImage(images.ui_score, center.x - (images.ui_score:getWidth() / 2), 0)
-  drawImage(images.ui_hiscore, 1280 - images.ui_hiscore:getWidth(), 100)
-  drawImage(images.logo, 50, GAME_HEIGHT - images.logo:getHeight() * 1.3)
-  
-  local score_text = string.format("%.0f M", score)
-  love.graphics.setFont(fonts.score)
-  love.graphics.printf(score_text, center.x - 135, 53, limit, "right")
-  
-  setColor("black")
-  score_text = string.format("%.0f M", hiscore)
-  love.graphics.setFont(fonts.high_score)
-  love.graphics.printf(score_text, 1075, 145, limit, "right")
-  setColor("white")
-  
-  if not isRunning then
-    if playerGotNewHighScore then
-      setColor("black")
-      love.graphics.printf("New High Score! " .. score_text, 0, center.y, GAME_WIDTH, "center")
-    end
-
-    -- Show info centered on the screen
-    local instructions = {
-      string.format("Press %s to start skiing", string.upper(controls.start)),
-      string.format("%s toggles fullscreen", string.upper(controls.toggle_fullscreen)),
-      string.format("%s quits the game", string.upper(controls.quit))
-    }
-    
-    local font_height = fonts.instructions:getHeight()
-    
-    setColor("light-blue accent-4", .3)
-    love.graphics.rectangle("fill", 0, center.y, GAME_WIDTH, font_height * (#instructions + 2))
-    setColor("white")
-    love.graphics.setFont(fonts.instructions)
-
-    for i = 1, #instructions do
-      love.graphics.printf(instructions[i], 0, center.y + (font_height * i), GAME_WIDTH, "center")
-    end
-  end
-end
-
-function level.draw(controls)
-  drawBackground()
-  drawForeground()
-  drawLanes()
-  drawObjects()
-  
+function level.drawPlayer()
   -- Draw whale according to game running state
   if not isRunning then
-    love.graphics.draw(images.player_idle, whale.x, whale.y, 0, 1, 1, whale.offsetX, whale.offsetY)
+    Graphics:drawIdlePlayer(whale.x, whale.y, whale.offsetX, whale.offsetY)
   else
     if (#level.objects == 0) and isRunning then
       whale.playAnimation()
     end
   end
-  
-  -- Draw user interface
-  drawGUI(controls)
-end
-
-function level.toggleFullscreen()
-  local isFullscreen = not love.window.getFullscreen()
-  love.window.setFullscreen(isFullscreen, "desktop")
-end
-
-function level.toggleMouseVisibility()
-  local state = not love.mouse.isVisible()
-  love.mouse.setVisible(state)
 end
 
 function level.startGame()
